+++ CHANGELOG.md
@@ -4,6 +4,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.21.2]
+
+### Added
+
+- Added `sub_registry_with_labels` method to `Registry`.
+  See [PR 145].
+- Added `with_labels` and `with_prefix_and_labels` constructors to `Registry`.
+  See [PR 147].
+
+[PR 145]: https://github.com/prometheus/client_rust/pull/145
+[PR 147]: https://github.com/prometheus/client_rust/pull/147
+
 ## [0.21.1]
 
 ### Added